AJ MacLeod:
I have been working for longer than I care to think on a model of an English
Electric Lightning (F.1A) for flightgear.

It has plenty of quirks and irregularities still (as did the original  :-)  but
I hope it should nevertheless be quite entertaining to fly.  The main caveat
is on the autopilot ILS modes which should currently be avoided (or much
better - fixed by someone who knows more about fgfs autopilots than I do!)

Any improvements anyone wishes to contribute are _very_ welcome - there is
plenty of scope for additions still.